fix: refactoring code in overlay adapter
