+++ src/contentscript/modules/index.ts
@@ -1,0 +1,5 @@
+import OverlayAdapter, { ManifestOverlayAdapter } from './adapter-overlay/src'
+
+export = {
+  'overlay-adapter.dapplet-base.eth': ManifestOverlayAdapter,
+}
