+++ src/contentscript/modules/state.ts
