+++ src/contentscript/overlay/root/components/DevModulesList/index.tsx
@@ -405,11 +405,7 @@ export const DevModule: FC<PropsDevModule> = (props) => {
 
           <div className={styles.blockInfo}>
             <h3 className={styles.dappletsTitle}>{mi.title}</h3>
-            {mi.isUnderConstruction ? //   onClick={() => { // <span
-            //     onDetailsClick(mi, vi)
-
-            //     setUnderConstructionDetails(true)
-            //     setModuleInfo(mi)
+            {mi.isUnderConstruction ? //     setModuleInfo(mi) //     setUnderConstructionDetails(true) //     onDetailsClick(mi, vi) //   onClick={() => { // <span
             //     setModuleVersion(vi)
             //   }}
             //   className={styles.dappletsSettingsIsUnderConstructionBlock}
